requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
qdrant-client>=1.12.0
sentence-transformers>=3.0.0
python-dotenv>=1.0.0
requests>=2.32.0
punq>=0.7.0
langchain-text-splitters>=0.3.0
pydantic>=2.0.0
pytest>=8.2.2
httpx>=0.27.0